pyproject.toml
[tool.poetry]
name = "django-oac"
version = "0.1.1"
description = ""
authors = ["Przemek Kaliś <przemek.kalis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
jwcrypto = "^0.7"
requests = "^2.24.0"
pendulum = "^2.1.2"
pyjwt = "^1.7.1"
cryptography = "^3.0"
django-ipware = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
pytest-django = "^3.9.0"
isort = "^5.4.2"
django_sslserver = "^0.22"
pytest-cov = "^2.10.1"
python-dotenv = "^0.14.0"
python-coveralls = "^2.9.3"
coverage = "<5"
flake8 = "^3.8.3"
toml = "^0.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
